Moved submission methods
